Move hoi commands into other cog
